Extract `Cache` and other types from `html` module
